switch chain working also deploying token
