+++ pages/index.tsx
@@ -35,11 +35,14 @@ export default function Home() {
 
   const { writeContract, data: hash, isPending } = useWriteContract();
 
-  const { data: dataConfirmed } = useWaitForTransactionReceipt({
+  const { data: dataConfirmed, } = useWaitForTransactionReceipt({
     hash: hash,
   });
 
   useEffect(() => {
+    if (!dataConfirmed) {
+      return;
+    }
     const response = dataConfirmed ? dataConfirmed.logs[4]?.topics : [];
     const beginIndex = 2;
     const endIndex = 26;
@@ -63,32 +66,32 @@ export default function Home() {
     setTransactionStatus("0");
   };
   const tokenAddress = erc20TokenAddress as `0x`;
-  const result = useReadContracts({ 
-    allowFailure: false, 
-    contracts: [ 
-      { 
-        address: tokenAddress, 
-        abi: erc20Abi, 
-        functionName: 'decimals', 
-      }, 
-      { 
-        address: tokenAddress, 
-        abi: erc20Abi, 
-        functionName: 'name', 
-      }, 
-      { 
-        address: tokenAddress, 
-        abi: erc20Abi, 
-        functionName: 'symbol', 
-      }, 
-      { 
-        address: tokenAddress, 
-        abi: erc20Abi, 
-        functionName: 'totalSupply', 
-      }, 
-    ] 
+  const result = useReadContracts({
+    allowFailure: false,
+    contracts: [
+      {
+        address: tokenAddress,
+        abi: erc20Abi,
+        functionName: 'decimals',
+      },
+      {
+        address: tokenAddress,
+        abi: erc20Abi,
+        functionName: 'name',
+      },
+      {
+        address: tokenAddress,
+        abi: erc20Abi,
+        functionName: 'symbol',
+      },
+      {
+        address: tokenAddress,
+        abi: erc20Abi,
+        functionName: 'totalSupply',
+      },
+    ]
   })
-  
+
 
   useEffect(() => {
     const contractAddress = getNetworkContract(chain?.id!);
@@ -113,7 +116,7 @@ export default function Home() {
   };
 
   return (
-    <>
+    <div>
       <Head>
         <title>Deploy Supertoken</title>
         <meta name="description" content="Generated by create next app" />
@@ -155,8 +158,7 @@ export default function Home() {
               </h1>
               {/* const { name, symbol, tokenAddress, setToken, deploySupertoken, tokenData, isLoading, supertoken } = props; */}
               <DeploySupertoken
-                //@ts-ignore
-                tokenAddress={erc20TokenAddress}
+                tokenAddress={erc20TokenAddress ?? ""}
                 name={name}
                 symbol={symbol}
                 setToken={setAddress}
@@ -166,9 +168,10 @@ export default function Home() {
                   symbol: symbol,
                   address: erc20TokenAddress,
                 }}
-                networkId={chainId}
+                
                 isLoading={isPending}
                 supertoken={supertoken}
+
               />
             </>
           ) : (
@@ -230,6 +233,6 @@ export default function Home() {
           </a>
         </div>
       </main>
-    </>
+    </div>
   );
 }
